Support various tx types in APIs
